mcap: flexible option argument positions
The mcap tool uses getopt to parse the command line options but hardcoded
the position of the option arguments. The order of the options is not
documented and will break when a user calls the program in another way.
It also complexifies the parsing and makes adding options harder.

Use optarg for simpler and flexible handling of the arguments of options
'-x', '-C' and '-p'.